+++ core/Updates/2.15.0-b16.php
@@ -13,7 +13,7 @@ use Piwik\Plugin\Manager;
 use Piwik\Updater;
 use Piwik\Updates;
 
-class Updates_2_15_0_b13 extends Updates
+class Updates_2_15_0_b16 extends Updates
 {
 
     public function doUpdate(Updater $updater)